requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
requests==2.31.0
python-dotenv==1.0.0
google-genai==1.0.0
jupyter==1.0.0
ipykernel==6.29.0
